Code refactoring
I refactored the code because the Cyclomatic Complexity was too high (20 according to IntelliJ CodeMetrics plugin). I created some extra methods to reduce the complexity, improve readability and make the quality of the code better. I think now the code for this class is easier to understand, the class complexity was reduced to 8.