Add files via upload
This file contains the annotations for the CFTR sequence, including important information such as promoter sequences and the most common mutated sequence (F508 del). I would suggest this be implemented, and then a practice may be run by taking a large sample sequence including the CFTR gene which I uploaded earlier, giving it a characteristic mutation, and then testing the project.

Signed-off-by: Karzinisierung <118777053+Karzinisierung@users.noreply.github.com>